+++ CHANGELOG.md
@@ -51,6 +51,7 @@ On the other hand, this also allowed us to deal with further issues that contain
 ### Fixed
 
 - Fixed a deviation from the paper, where `DeltaFScore` also took into account the agreement in predictions of the negative label. ([#51](https://github.com/webis-de/small-text/pull/51))
+- Fixed a bug in `KappaAverage` that affected the stopping behavior. ([#52](https://github.com/webis-de/small-text/pull/52))
 
 ---